Add buslogger class and setup for screen
Converted an old script to a class to retrieve and log the bus
information from the Dublin Bus api. A utility file was added to collect
useful functions, the first is a function that returns the root
directory of the project. The file that will contain the class for
interacting with the LCD screen was also added.