Added main script file and centered flag
The first functional version of the program has been added. The program
now is nearly feature complete as far as reporting the bus times to the
screen and saving the data to date specific files. The program will be
scheduled using cron to run from 6am when buses start till about
midnight when the last bus leaves. In the next major version, the
program should handle it's own scheduling as well as handle data
requests asynchronously. Also in this commit some subtle changes have
been made to buslogger to handle errors a little better and to
screenwriter which now has a flag that will automatically center a given
string if a flag is set.