Change function file name and added tests
Changed the name of a function file from utils.py to helpers.py as
utils.py overlaps with the module of the same name. This meant
redefining its import across a few files. The block of code that checks
if a file exists, then opens it and appends some data has been moved its
own function in helpers.py as similar functionality was needed elsewhere
in the program and I'm trying to stay DRY. Tests have also been added
for the helper function file ensuring near complete test coverage.
Finally a test was added for the center flag functionality in
screenwriter that was added in the previous commit.