Implemented schedule and multi-threading in main
Since the rudimentary scheduling was not enough to show the bus times
correctly a new system has been implemented. The program now uses the
schedule python module and multi-threading to update the data and change
between bus times on the screen at set intervals on separate threads.
This should mean that the bus times will be more consistent on the
screen.